+++ backend/src/index.ts
@@ -20,9 +20,9 @@ import { dbInit } from './database/index';
 
 dbInit();
 
-import {router as LessonController} from './lesson/api/controller';
-import {router as SubjectController} from './lesson/api/controller';
-import {router as VirtualEntityController} from './lesson/api/controller';
+import {router as LessonController} from './lesson/api/lessonController';
+import {router as SubjectController} from './subject/api/subjectController';
+import {router as VirtualEntityController} from './virtualEntity/api/virtualEntityController';
 
 const PORT = process.env.PORT || 8080;
 
@@ -32,7 +32,6 @@ app.use(express.json());
 app.use(cors());
 app.use('/lesson', LessonController)
 app.use('/subject', SubjectController)
-
 app.use('/virtualEntity', VirtualEntityController)
 
 app.get('/', (req, res) => {
