+++ backend/src/virtualEntity/service/virtualEntityService.ts
